requirements.txt
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-yasg==1.21.7
pymongo==4.8.0
datetime==5.5
djangorestframework-simplejwt==5.3.1
requests==2.32.3
pyotp==2.9.0
qrcode==7.4.2
python-magic==0.4.27
pillow==10.4.0
django-filter==24.3
openai==1.40.2
langchain==0.2.12
langchain-community==0.2.11
tiktoken==0.7.0
langchain_openai==0.1.20
pypdf==4.3.1